.opex_final_anexoA/requirements.txt
google-api-python-client>=2.108.0
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
python-docx>=1.1.0
